# openapispec.yaml
openapi: 3.0.3
...
info:
  title: fwe_backend
  version: 0.0.1
  contact:
    name: Nils Heißner
    email: nils.heissner@stud.h-da.de
    url: h-da.de
  description: The API for the Learning Corner project
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
servers:
  - url: 'http://localhost:4000'
paths:
...
  '/users/{userID}/course/{courseID}':
    parameters:
      - schema:
          type: string
        name: userID
        in: path
        required: true
      - schema:
          type: string
        name: courseID
        in: path
        required: true
    put:
      summary: ''
      operationId: put-users-userID-course-courseID
      responses:
        '200':
          description: OK
      description: ''
    delete:
      summary: ''
      operationId: delete-users-userID-course-courseID
      responses:
        '204':
          description: No Content
  /courses:
    get:
      description: Get courses
      operationId: get-courses
      parameters:
        - in: query
          name: name
          schema:
            type: string
          description: The name to filter courses by
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
      security:
        - Bearer: []
    post:
      description: Post course
      operationId: post-courses
      requestBody:
        description: The Course Object to create and the newly added Files
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Error while creating
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        '409':
          description: Error while creating - Course name already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: string
      security:
        - Bearer: []
  '/courses/{id}':
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: The id of the course to work with
    get:
      description: Get an Course by Id
      operationId: get-courses-id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          description: Invalid Id
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
      security:
        - Bearer: []
    put:
      summary: update a course by only course creator
      operationId: put-course
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
      security:
        - Bearer: []
    delete:
      summary: delete a course by only course creator
      operationId: delete-courses-courseID-users-userID
      responses:
        '204':
          description: No Content
      security:
        - Bearer: []
  '/courses/{id}/section':
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: The id of the course
    post:
      description: Create section in course
      operationId: post-section-course
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
        '400':
          description: Error while creating
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        '401':
          description: Error while creating - not a creator of the course
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: string
        '404':
          description: Invalid Id
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
      requestBody:
        description: The Section Object to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Section'
      security:
        - Bearer: []
  '/courses/{courseId}/section/{sectionId}':
    parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: string
          format: uuid
        description: The id of the course
      - in: path
        name: sectionId
        required: true
        schema:
          type: string
          format: uuid
        description: The id of the section
    put:
      description: Update section in a course
      operationId: put-section-course
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
        '400':
          description: Error while creating
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        '401':
          description: Error while creating - not a creator of the course
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: string
        '404':
          description: Invalid Id
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Section'
      security:
        - Bearer: []
    delete:
      description: Delete section in a course
      operationId: delete-section-course
      responses:
        '204':
          description: No Content
        '400':
          description: Error while creating
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        '401':
          description: Error while creating - not a creator of the course
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: string
        '404':
          description: Invalid Id
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
      security:
        - Bearer: []
...
components:
  schemas:
    User:
      title: User
      type: object
      description: The User Objet
      examples:
        - id: d46eb448-768b-4e74-a7dd-60cd9e7ee8a2
          firstname: Max
          lastname: Mustermann
          email: user@learning-corner.de
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the given user.
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        photo:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        courses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        joinedCourses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
      required:
        - id
        - firstName
        - lastName
        - email
        - photo
    Course:
      title: Course
      type: object
      description: A Course
      examples:
        - id: d46eb448-768b-4e74-a7dd-60cd9e7ee8a2
          name: Web-Development for Dummies
          description: Learn the 101 of Web-Development
          lecturer:
            $ref: '#/components/schemas/User'
          participants:
            type: array
            items:
              type:
                $ref: '#/components/schemas/User'
          groups:
            type: array
            items:
              type:
                $ref: '#/components/schemas/Group'
          files:
            type: array
            items:
              type:
                $ref: '#/components/schemas/File'
          quizzes:
            type: array
            items:
              type:
                $ref: '#/components/schemas/Quiz'
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        lecturer:
          type:
            $ref: '#/components/schemas/User'
          $ref: '#/components/schemas/User'
        participants:
          type: array
          items:
            type:
              $ref: '#/components/schemas/User'
            $ref: '#/components/schemas/User'
        groups:
          type: array
          items:
            type:
              $ref: '#/components/schemas/Group'
            $ref: '#/components/schemas/Group'
        sections:
          type: array
          items:
            type:
              $ref: '#/components/schemas/Section'
            $ref: '#/components/schemas/Section'
        quizzes:
          type: array
          items:
            type:
              $ref: '#/components/schemas/Quiz'
            $ref: '#/components/schemas/Quiz'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - description
        - lecturer
    File:
      title: File
      type: object
      description: A File
      examples:
        - id: d46eb448-768b-4e74-a7dd-60cd9e7ee8a2
          name: folien.pdf
          course:
            type:
              $ref: '#/components/schemas/Course'
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        content:
          type: string
        owner:
          oneOf:
            - $ref: '#/components/schemas/Course'
            - $ref: '#/components/schemas/Group'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - course
      x-stoplight:
        id: c00d5a960f6ee
    Section:
      title: Section
      description: A content section in a course/group
      type: object
      x-stoplight:
        id: 8703e8cdbf634
      properties:
        id:
          type: string
          format: uuid
        heading:
          type: string
        description:
          type: string
        text:
          type: string
        course:
          type:
            $ref: '#/components/schemas/Course'
          $ref: '#/components/schemas/Course'
        group:
          type:
            $ref: '#/components/schemas/Group'
          $ref: '#/components/schemas/Group'
        files:
          type: array
          items:
            type:
              $ref: '#/components/schemas/File'
            $ref: '#/components/schemas/File'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - heading
        - description
        - text
    Group:
      title: Group
      x-stoplight:
        id: hincqq4kpsm12
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        course:
          $ref: '#/components/schemas/Course'
        members:
          type: array
          items:
            type:
              $ref: '#/components/schemas/User'
            $ref: '#/components/schemas/User'
        sections:
          type: array
          items:
            type:
              $ref: '#/components/schemas/Section'
            $ref: '#/components/schemas/Section'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Quiz:
      title: Quiz
      type: object
      description: A Quiz
      examples:
        - id: d46eb448-768b-4e74-a7dd-60cd9e7ee8a2
          name: HTML Quiz - Beginner
          description: Check your basic knowledge about HTML
          questions:
            type: array
            items:
              type:
                $ref: '#/components/schemas/QuizQuestion'
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            type:
              $ref: '#/components/schemas/QuizQuestion'
            $ref: '#/components/schemas/QuizQuestion'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - description
        - questions
    QuizQuestion:
      title: QuizQuestion
      type: object
      description: A QuizQuestion
      examples:
        - id: d46eb448-768b-4e74-a7dd-60cd9e7ee8a2
          text: Is HTLM > JS?
          correctAnswer: 0
          answers:
            type: array
            items:
              type:
                $ref: '#/components/schemas/QuizQuestionAnswer'
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        correctAnswer:
          type: number
        answers:
          type: array
          items:
            type:
              $ref: '#/components/schemas/QuizQuestionAnswer'
            $ref: '#/components/schemas/QuizQuestionAnswer'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - text
        - correctAnswer
        - answers
    QuizQuestionAnswer:
      title: QuizQuestionAnswer
      type: object
      description: A QuizQuestionAnswer
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - text
      examples:
        - id: d46eb448-768b-4e74-a7dd-60cd9e7ee8a2
          text: 'No'
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
...